Added the drawing stuff for the second user.
Moved plotChart into a separate procedure, this either draws each user's
curves or clear series if neither user is selected.

Added a new procedure setToday, to draw a vertical bar to indicate today, on the chart.
It looks at each user series to try and determine the range.
The secondary combined (if used) has the greatest range.
If both user's combines are displayed, use the min and max of the two.
If neither combined is used, use defaults..